+++ src/Avalonia.Controls.WebView.Core/Browser/BrowserIFrameAdapter.cs
@@ -67,6 +67,12 @@ internal class BrowserIFrameAdapter : JSObjectControlHandle, IWebViewAdapter,
         }
     }
 
+    public string? UserAgent
+    {
+        get => WebViewInterop.GetNavigatorUserAgent();
+        set { }
+    }
+
     public void SizeChanged(PixelSize containerSize) { }
 
     public void SetParent(IPlatformHandle parent) { }