+++ src/Avalonia.Controls.WebView/CHANGELOG.md
@@ -3,6 +3,7 @@
 - Compatibility with Avalonia 12.0, while keeping 11.1 support
 - Dropped .NET Standard and .NET Framework support
 - NativeWebView and NativeWebDialog now support Browser backend, but with sandboxing limitations
+- NativeWebView and NativeWebDialog now have UserAgent property
 
 # 11.3.16